@@ -1,16 +1,22 @@
 import Cookies from 'universal-cookie';
-import { selectedProjectCookieName, defaultProjectId } from './Utilities/SelectedProject';
+import {
+  selectedProjectCookieName,
+  defaultProjectId,
+} from './Utilities/SelectedProject';
 import packageInfo from '../package.json';
 
 // Kontent.ai
-import { camelCasePropertyNameResolver, DeliveryClient } from '@kontent-ai/delivery-sdk';
+import {
+  camelCasePropertyNameResolver,
+  DeliveryClient,
+} from '@kontent-ai/delivery-sdk';
 
-const projectId = process.env.VUE_APP_PROJECT_ID || '';
-const previewApiKey = process.env.VUE_APP_PREVIEW_API_KEY || '';
+const projectId = import.meta.env.VUE_APP_PROJECT_ID || '';
+const previewApiKey = import.meta.env.VUE_APP_PREVIEW_API_KEY || '';
 
 const sourceTrackingHeaderName = 'X-KC-SOURCE';
 
-const cookies = new Cookies(document.cookies);
+const cookies = new Cookies(document.cookie);
 let currentProjectId = projectId || cookies.get(selectedProjectCookieName);
 if (currentProjectId) {
   cookies.set(selectedProjectCookieName, currentProjectId, { path: '/' });
@@ -24,8 +30,8 @@ let Client = new DeliveryClient({
   projectId: currentProjectId,
   propertyNameResolver: camelCasePropertyNameResolver,
   previewApiKey: previewApiKey,
-  globalQueryConfig: {
-    usePreviewMode: isPreview()
+  defaultQueryConfig: {
+    usePreviewMode: isPreview(),
   },
   globalHeaders: () => [
     {
@@ -35,14 +41,13 @@ let Client = new DeliveryClient({
   ],
 });
 
-
-const resetClient = (newProjectId) => {
+const resetClient = (newProjectId: string) => {
   Client = new DeliveryClient({
     projectId: newProjectId,
     propertyNameResolver: camelCasePropertyNameResolver,
     previewApiKey: previewApiKey,
-    globalQueryConfig: {
-      usePreviewMode: isPreview()
+    defaultQueryConfig: {
+      usePreviewMode: isPreview(),
     },
     globalHeaders: () => [
       {
@@ -51,11 +56,8 @@ const resetClient = (newProjectId) => {
       },
     ],
   });
-  const cookies = new Cookies(document.cookies);
+  const cookies = new Cookies(document.cookie);
   cookies.set(selectedProjectCookieName, newProjectId, { path: '/' });
-}
-
-export {
-  Client,
-  resetClient
 };
+
+export { Client, resetClient };
