@@ -1,95 +1,92 @@
 <template>
-    <div
-        class="container"
-        v-if="!brewer"
-    >
-    </div>
-    <div
-        class="container"
-        v-else
-    >
-        <article class="product-detail">
-            <div class="row">
-                <div class="col-md-12">
-                    <header>
-                        <h2>{{name}}</h2>
-                    </header>
-                </div>
-            </div>
-            <div class="row-fluid">
-                <div class="col-lg-7 col-md-6">
-                    <figure class="image">
-                        <img
-                            :alt="name"
-                            class=""
-                            :src="imageLink"
-                            :title="name"
-                        />
-                    </figure>
-                    <div class="description">
-                        <RichTextElement
-                            v-if="descriptionElement"
-                            :element="descriptionElement"
-                        />
-                    </div>
-                </div>
-            </div>
-        </article>
-    </div>
+  <div class="container" v-if="!brewer"></div>
+  <div class="container" v-else>
+    <article class="product-detail">
+      <div class="row">
+        <div class="col-md-12">
+          <header>
+            <h2>{{ data.name }}</h2>
+          </header>
+        </div>
+      </div>
+      <div class="row-fluid">
+        <div class="col-lg-7 col-md-6">
+          <figure class="image">
+            <img
+              :alt="data.name"
+              class=""
+              :src="data.imageLink"
+              :title="data.name"
+            />
+          </figure>
+          <div class="description">
+            <RichTextElement
+              v-if="data.descriptionElement"
+              :element="data.descriptionElement"
+            />
+          </div>
+        </div>
+      </div>
+    </article>
+  </div>
 </template>
 
-<script>
+<script setup lang="ts">
 import { Client } from '../Client.js';
-import RichTextElement from './RichTextElement.vue'
+import RichTextElement from './RichTextElement.vue';
 import { resolveChangeLanguageLink } from '../Utilities/RouterLink';
+import { useI18n } from 'vue-i18n';
+import { onMounted, ref, watch } from 'vue';
+import { computed } from '@vue/reactivity';
+import { useRoute, useRouter } from 'vue-router';
+import type { Brewer } from '@/models';
+import type { Elements } from '@kontent-ai/delivery-sdk';
 
-export default {
-  name: 'Brewer',
-  props: ['language'],
-  data: () => ({
-    brewer: null,
-    name: '',
-    imageLink: '',
-    descriptionElement: null,
-  }),
-  watch: {
-    brewer: function(newBrewer){
-      if(!newBrewer){
-        return;
-      }
-      this.name = newBrewer.elements.productName.value;
-      this.imageLink = newBrewer.elements.image.value[0].url;
-      this.descriptionElement = newBrewer.elements.longDescription;
-    },
-    language: function(){
-      this.fetchBrewer();
-    }
-  },
-  methods:{
-    fetchBrewer: function () {
-      var query = Client.items()
-        .type('brewer')
-        .equalsFilter('url_pattern', this.$route.params.brewerSlug)
-
-      if(this.language){
-        query.languageParameter(this.language)
-      }
-      query
-        .toPromise()
-        .then(response => {
-          this.brewer = response.data.items[0]
-          
-          if(this.brewer.system.language !== this.language) {
-            this.$router.replace({path: resolveChangeLanguageLink(this.$route.path, this.brewer.system.language)})
-          }
-        })
-    }
-  },
-  mounted: function(){
-    this.fetchBrewer();
-  },
-  components: {
-    RichTextElement,
-  },
+interface BrewerData {
+  name: string;
+  imageLink: string;
+  descriptionElement: Elements.RichTextElement | undefined;
 }
+
+const { locale } = useI18n();
+const route = useRoute();
+const router = useRouter();
+
+const brewer = ref<Brewer | null>(null);
+const data = computed<BrewerData>(() => ({
+  name: brewer.value?.elements.productName.value ?? '',
+  imageLink: brewer.value?.elements.image.value[0].url ?? '',
+  descriptionElement: brewer.value?.elements.longDescription,
+}));
+
+const fetchBrewer = () => {
+  var query = Client.items<Brewer>()
+    .type('brewer')
+    .equalsFilter('url_pattern', route.params.brewerSlug as string);
+
+  if (locale.value) {
+    query.languageParameter(locale.value);
+  }
+
+  query.toPromise().then((response) => {
+    brewer.value = response.data.items[0];
+
+    if (brewer.value.system.language !== locale.value) {
+      router.replace({
+        path: resolveChangeLanguageLink(
+          route.path,
+          brewer.value.system.language
+        ),
+      });
+    }
+  });
+};
+
+onMounted(() => {
+  fetchBrewer();
+});
+
+watch(locale, () => {
+  fetchBrewer();
+});
 </script>
