@@ -1,94 +1,86 @@
 <template>
-    <div class="product-page row">
-        <div class="flex">
-            <CoffeeFilter 
-                :language="this.language"
-                :processings="processings"
-                :productStatuses="productStatuses"
-                :filter="filter"
-                @set-filter="setFilter"
-            />
-            <Coffees 
-                :language="this.language"
-                :coffees="coffees"
-                :filter="filter"
-            />
-        </div>
+  <div class="product-page row">
+    <div class="flex">
+      <CoffeeFilter
+        :processings="processings"
+        :productStatuses="productStatuses"
+        :filter="filter"
+        @set-filter="setFilter"
+      />
+      <Coffees :coffees="coffees" :filter="filter" />
     </div>
+  </div>
 </template>
 
-<script>
-import { Filter } from '../Utilities/CoffeeFilter'
-import CoffeeFilter from './CoffeeFilter.vue'
-import Coffees from './Coffees.vue'
+<script setup lang="ts">
+import { Filter } from '../Utilities/CoffeeFilter';
+import CoffeeFilter from './CoffeeFilter.vue';
+import Coffees from './Coffees.vue';
 import { Client } from '../Client.js';
-import { initLanguageCodeObject, defaultLanguage } from '../Utilities/LanguageCodes';
+import {
+  defaultLanguage,
+  initLanguageCodeObjectWithArray,
+} from '../Utilities/LanguageCodes';
+import { useI18n } from 'vue-i18n';
+import { onMounted, ref, watch } from 'vue';
+import type { Coffee } from '@/models';
+import type { ITaxonomyTerms } from '@kontent-ai/delivery-sdk';
 
-export default {
-  name: 'CoffeeStore',
-  props: ['language'],
-  data: () => ({
-    coffees: [],
-    processings: [],
-    productStatuses: [],
-    filter: new Filter()
-  }),
-  components: {
-    CoffeeFilter,
-    Coffees
-  },
-  methods: {
-    fetchData: function(language) {
-      const coffeesList = initLanguageCodeObject();
-      var query = Client.items()
-        .type('coffee')
-        .orderParameter('elements.product_name', 'desc');
+const { locale } = useI18n();
+const coffees = ref<Array<Coffee>>([]);
+const processings = ref<Array<ITaxonomyTerms>>([]);
+const productStatuses = ref<Array<ITaxonomyTerms>>([]);
+const filter = ref<Filter>(new Filter());
 
-      if (language) {
-        query.languageParameter(language);
-      }
+const fetchData = (language: string) => {
+  const coffeesList = initLanguageCodeObjectWithArray<Coffee>();
+  var query = Client.items<Coffee>()
+    .type('coffee')
+    .orderParameter('elements.product_name', 'desc');
 
-      query.toPromise()
-        .then(response => {
-          if (language) {
-            coffeesList[language] = response.data.items;
-          } else {
-            coffeesList[defaultLanguage] = response.data.items;
-          }
-          this.coffees = coffeesList[language];
-        });
-    },
-    fetchProcessings: function() {
-      Client.taxonomy('processing')
-        .toPromise()
-        .then(response => {
-          this.processings = response.data.taxonomy.terms;
-        });     
-    },
-    fetchProductStatuses: function() {
-      Client.taxonomy('product_status')
-        .toPromise()
-        .then(response => {
-          this.productStatuses = response.data.taxonomy.terms;
-        });
-    },
-    setFilter: function (newFilter) {
-      this.filter = newFilter
-    },
-  },
-  watch: {
-    language: function () {
-      this.fetchData(this.language);
-      this.fetchProcessings();
-      this.fetchProductStatuses();
-    }
-  },
-  mounted: function() {
-    this.fetchData(this.language);
-    this.fetchProcessings();
-    this.fetchProductStatuses();
+  if (language) {
+    query.languageParameter(language);
   }
 
-}
-</script>
+  query.toPromise().then((response) => {
+    if (language) {
+      coffeesList[language] = response.data.items;
+    } else {
+      coffeesList[defaultLanguage] = response.data.items;
+    }
+    coffees.value = coffeesList[language];
+  });
+};
 
+const fetchProcessings = (): void => {
+  Client.taxonomy('processing')
+    .toPromise()
+    .then((response) => {
+      processings.value = response.data.taxonomy.terms;
+    });
+};
+
+const fetchProductStatuses = (): void => {
+  Client.taxonomy('product_status')
+    .toPromise()
+    .then((response) => {
+      productStatuses.value = response.data.taxonomy.terms;
+    });
+};
+
+const setFilter = (newFilter: Filter): void => {
+  filter.value = newFilter;
+};
+
+watch(locale, () => {
+  fetchData(locale.value);
+  fetchProcessings();
+  fetchProductStatuses();
+});
+
+onMounted(() => {
+  fetchData(locale.value);
+  fetchProcessings();
+  fetchProductStatuses();
+});
+</script>
