@@ -1,11 +1,10 @@
 <template>
-    <div class="container">
-        <div :style="{'text-align': 'center'}">
-            <h1>404</h1>
-            <h2>{{$t('NotFound.message')}}</h2>
-        </div>  
+  <div class="container">
+    <div :style="{ 'text-align': 'center' }">
+      <h1>404</h1>
+      <h2>{{ $t('NotFound.message') }}</h2>
     </div>
+  </div>
 </template>
 
-<script>
-</script>
+<script setup lang="ts"></script>
