@@ -1,26 +1,22 @@
 <template>
-    <div class="row">
-        <h1 class="title-tab">{{$t('OurStory.title')}}</h1>
-        <div class="col-sm-12">
-            <div 
-                class="ourstory-section center-text" 
-                v-bind:style='sectionStyleObject'
-            >
-                {{$t('OurStory.text')}}
-            </div>
-        </div>
+  <div class="row">
+    <h1 class="title-tab">{{ t('OurStory.title') }}</h1>
+    <div class="col-sm-12">
+      <div
+        class="ourstory-section center-text"
+        :style="sectionStyleObject"
+      >
+        {{ t('OurStory.text') }}
+      </div>
     </div>
+  </div>
 </template>
 
-<script>
+<script setup lang="ts">
+import { useI18n } from 'vue-i18n';
 import StoryImage from '../Images/our-story.jpg';
 
-export default {
-  name: 'our-story',
-  data: () => ({
-    sectionStyleObject: {
-      backgroundImage: 'url(' + StoryImage + ')',
-    },
-  }),
-}
+const { t } = useI18n();
+
+const sectionStyleObject = { backgroundImage: `url(${StoryImage})` };
 </script>
