@@ -1,13 +1,12 @@
 <template>
-    <div class="spinner-box">
-        <span class="spinner"></span>
-        <span class="spinner-message">{{message}}</span>
-    </div>
+  <div class="spinner-box">
+    <span class="spinner"></span>
+    <span class="spinner-message">{{ props.message }}</span>
+  </div>
 </template>
 
-<script>
-export default {
-  name: 'SpinnerBox',
-  props:['message']
-}
+<script setup lang="ts">
+const props = defineProps<{
+  message: string;
+}>();
 </script>
